requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
bitstring==3.1.7
certifi==2021.5.30
cffi==1.14.5
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.7
django-admin-list-filter-dropdown==1.0.3
django-admincharts==0.1.2
django-allauth==0.43.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==3.0.1
django-bower==5.2.0
django-chartjs==2.2.1
django-crispy-forms==1.12.0
django-debug-toolbar==3.2
django-filter==2.4.0
django-jsonfield==1.4.1
django-nvd3==0.9.7
django-related-admin==1.0.1
django-tables2==2.4.0
djangorestframework==3.12.4
djcacheutils==3.0.0
ecdsa==0.16.1
esptool==3.0
filelock==3.0.12
idna==3.2
Jinja2==3.0.2
Markdown==3.3.4
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.3.2
platformdirs==2.3.0
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
pyserial==3.5
python-dateutil==2.8.2
python-memcached==1.59
python-nvd3==0.14.2
python-slugify==1.1.4
python3-openid==3.2.0
pytz==2021.1
reedsolo==1.5.4
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
tenacity==8.0.1
Unidecode==1.3.2
urllib3==1.26.6
virtualenv==20.7.2
whitenoise==5.2.0
